[JENKINS-73119] Make folders dependency required, not optional
If the folders plugin is not installed, jobs fail to start and report:
"Exception evaluating if the gueue can run the task"

Once the folders plugin is declared as a required dependency rather than
an optional dependency, then the folders plugin will be installed when
the plugin is updated and jobs will run successfully.

Confirmed the failure by running Jenkins 2.426.3 with the build blocker
plugin 165.v5ecb_fb_f61520 installed.  I created a freestyle job and
ran it.  The run failed with the message "Exception evaluating if the
gueue can run the task".

Confirmed the failure is resolved by installing the folders plugin
6.858.v898218f3609d.  That is the version offered by the plugin bill of
materials version that is defined in this pom file.