[JENKINS-73119] Catch the class not found exception
If the folders plugin is not installed, jobs fail to start and report:
"Exception evaluating if the gueue can run the task"

Once the class not found exception is caught, then the job runs as expected.

Confirmed the failure by running Jenkins 2.426.3 with the build blocker
plugin 165.v5ecb_fb_f61520 installed.  I created a freestyle job and
ran it.  The run failed with the message "Exception evaluating if the
gueue can run the task".

Confirmed the failure is resolved by installing the build of this commit
and ran the same job without error.  Confirmed that the class not found
exception was written to the Jenkins logger when I enabled FINE logging
for the BuildBlockerQueueTaskDispatcher.  FINE logging is disabled by
default, so the Jenkins log will not be cluttered by those exception
messages unless an administrator specifically enables it.